requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
networkx>=3.0
simpy>=4.0.0
pyyaml>=6.0
pytest>=7.0.0
pandas>=2.0.0
